# Draw splats as tight oriented quads, instead of conservative screen aligned squares