--- a/Assets/GaussianSplatting/Scripts/GaussianSplatRenderer.cs
+++ b/Assets/GaussianSplatting/Scripts/GaussianSplatRenderer.cs
@@ -12,8 +12,9 @@ using UnityEngine.Rendering;
 class GaussianSplatRenderSystem
 {
     static ProfilerMarker s_ProfDraw = new(ProfilerCategory.Render, "GaussianSplat.Draw", MarkerFlags.SampleGPU);
-    internal static ProfilerMarker s_ProfCompose = new(ProfilerCategory.Render, "GaussianSplat.Compose", MarkerFlags.SampleGPU);
-    
+    static ProfilerMarker s_ProfCompose = new(ProfilerCategory.Render, "GaussianSplat.Compose", MarkerFlags.SampleGPU);
+    static ProfilerMarker s_ProfCalcView = new(ProfilerCategory.Render, "GaussianSplat.CalcView", MarkerFlags.SampleGPU);
+
     public static GaussianSplatRenderSystem instance => ms_Instance ??= new GaussianSplatRenderSystem();
     static GaussianSplatRenderSystem ms_Instance;
     
@@ -134,7 +135,9 @@ class GaussianSplatRenderSystem
             mpb.SetInteger("_DisplayIndex", gs.m_RenderMode == GaussianSplatRenderer.RenderMode.DebugPointIndices ? 1 : 0);
             mpb.SetInteger("_DisplayChunks", gs.m_RenderMode == GaussianSplatRenderer.RenderMode.DebugChunkBounds ? 1 : 0);
 
+            cmb.BeginSample(s_ProfCalcView);
             gs.CalcViewData(cmb, cam, matrix);
+            cmb.EndSample(s_ProfCalcView);
 
             // draw
             int indexCount = 6;
@@ -261,7 +264,6 @@ public class GaussianSplatRenderer : MonoBehaviour
     Hash128 m_PrevHash;
 
     static readonly ProfilerMarker s_ProfSort = new(ProfilerCategory.Render, "GaussianSplat.Sort", MarkerFlags.SampleGPU);
-    static readonly ProfilerMarker s_ProfView = new(ProfilerCategory.Render, "GaussianSplat.View", MarkerFlags.SampleGPU);
 
     [field: NonSerialized] public bool editModified { get; private set; }
     [field: NonSerialized] public uint editSelectedSplats { get; private set; }
@@ -439,8 +441,6 @@ public class GaussianSplatRenderer : MonoBehaviour
         if (cam.cameraType == CameraType.Preview)
             return;
 
-        using var prof = s_ProfView.Auto();
-
         var tr = transform;
 
         Matrix4x4 matView = cam.worldToCameraMatrix;
@@ -471,7 +471,7 @@ public class GaussianSplatRenderer : MonoBehaviour
         cmb.SetComputeFloatParam(m_CSSplatUtilities, "_SplatOpacityScale", m_OpacityScale);
         cmb.SetComputeIntParam(m_CSSplatUtilities, "_SHOrder", m_SHOrder);
 
-        m_CSSplatUtilities.GetKernelThreadGroupSizes((int)KernelIndices.CalcViewData, out uint gsX, out uint gsY, out uint gsZ);
+        m_CSSplatUtilities.GetKernelThreadGroupSizes((int)KernelIndices.CalcViewData, out uint gsX, out _, out _);
         cmb.DispatchCompute(m_CSSplatUtilities, (int)KernelIndices.CalcViewData, (m_GpuView.count + (int)gsX - 1)/(int)gsX, 1, 1);
     }