--- a/Assets/GaussianSplatting/Shaders/SplatUtilities.compute
+++ b/Assets/GaussianSplatting/Shaders/SplatUtilities.compute
@@ -115,18 +115,15 @@ void CSCalcViewData (uint3 id : SV_DispatchThreadID)
         cov3d1 *= _SplatScale * _SplatScale;
         float3 cov2d = CalcCovariance2D(splat.pos, cov3d0, cov3d1, _MatrixMV, _MatrixP, _VecScreenParams);
         
-        float det = cov2d.x * cov2d.z - cov2d.y * cov2d.y;
-
-        // Guard against precision issues that would result in occasional splat explosions, where a single splat would explode to covering the whole screen at rare angles / positions as its covariance matrix becomes degenerate.
-        det = max(1e-11, det);
-
-        float mid = 0.5f * (cov2d.x + cov2d.z);
-        float lambda1 = mid + sqrt(max(0.1f, mid * mid - det));
-        float lambda2 = mid - sqrt(max(0.1f, mid * mid - det));
-        float radius = ceil(3.f * sqrt(max(lambda1, lambda2)));
-
-        float3 conic = float3(cov2d.z, -cov2d.y, cov2d.x) * rcp(det);
-        view.conicRadius = float4(conic, radius);
+        float diag1 = cov2d.x, diag2 = cov2d.z, offDiag = cov2d.y;
+        float mid = 0.5f * (diag1 + diag2);
+        float radius = length(float2((diag1 - diag2) / 2.0, offDiag));
+        float lambda1 = mid + radius;
+        float lambda2 = max(mid - radius, 0.1);        
+        float2 diagVec = normalize(float2(offDiag, lambda1 - diag1));
+        diagVec.y = -diagVec.y;
+        view.axis1 = min(sqrt(2.0 * lambda1), 1024.0) * diagVec;
+        view.axis2 = min(sqrt(2.0 * lambda2), 1024.0) * float2(diagVec.y, -diagVec.x);        
 
         float3 worldViewDir = _VecWorldSpaceCameraPos.xyz - centerWorldPos;
         float3 objViewDir = mul((float3x3)_MatrixWorldToObject, worldViewDir);