… Doc/RefImages/SBIR_bicycle0.png
… Doc/RefImages/SBIR_garden30.png
… Doc/RefImages/SBIR_truck30.png