Using the same quad orientation math as [antimatter15/splat](https://github.com/antimatter15/splat). Quite a bit faster! Bicycle scene:
- PC DX12, RTX 3080Ti: 12.6ms -> 8.1ms
- Mac, M1 Max: 31.8ms -> 23.6ms